Refactor: Resolve all warnings

+++ TEMU/pytemu/src/instructions/r_type_ins/div_ins.py
@@ -9,7 +9,7 @@ class div_ins(RIns.R_Ins):
         super().__init__(instruction)
 
     def execute(self, cpu):
-        lo = c_int32(cpu[self._rs].low32).value / c_int32(cpu[self._rt].low32).value
+        lo = int(c_int32(cpu[self._rs].low32).value / c_int32(cpu[self._rt].low32).value)
         hi = c_int32(cpu[self._rs].low32).value % c_int32(cpu[self._rt].low32).value
         cpu[34].low32 = c_int32(lo).value
         cpu[33].low32 = c_int32(hi).value
